module for opt in persister
